Creación del useContext para favoritos, y utilización del mismo en petCard
