+++ components/petCard/pet-card.tsx
@@ -10,24 +10,26 @@ import Link from "next/link";
 import { useAuth } from "@/contexts/auth-context";
 import { addFavorite, deleteFavorite, getFavorites } from "@/utils/favorites-posts.http";
 import { Alert } from "@material-tailwind/react";
+import { useFavorites } from "@/contexts/favorites-context";
+import { Favorites } from "@/types/favorites";
 
 //Defini estos tipos para que el componente no tenga errores, esto debera cambiar en el futuro cuando el endpoint que conecta
 //posts con pets este implementado
-
-
-
 type PetCardProps = {
     post: any;
     className?: string
     isPost?: boolean;
 };
 
-
 export default function PetCard({ post, className, isPost }: PetCardProps) {
-    const [isFavorite, setIsFavorite] = useState<boolean>(false);
+    const { favorites, setFavorites } = useFavorites(); // Usamos el contexto
     const [successMessage, setSuccessMessage] = useState("");
     const [errorMessage, setErrorMessage] = useState("");
     const { authToken } = useAuth(); // Hook de autenticación
+
+    // Determinar si el post está en favoritos con una sola evaluación
+    const isFavorite = favorites.some((fav: Favorites) => fav.postId === (post as Post).id);
+
     // Leer el estado del 'localStorage' (si existe) al cargar el componente
     // const [isFavorite, setIsFavorite] = useState<boolean>(() => {
     //     // Intentamos leer el valor del 'localStorage' usando el id del post
@@ -45,21 +47,6 @@ export default function PetCard({ post, className, isPost }: PetCardProps) {
     // };
     // console.log(post)
     // Cargar favoritos al montar el componente
-    useEffect(() => {
-        if (!authToken) return;
-        
-        const fetchFavorites = async () => {
-            try {
-                const data = await getFavorites(authToken);
-                // Verificamos si el post actual está en los favoritos del usuario
-                setIsFavorite(data.some((fav: any) => fav.postId === post.id));
-            } catch (error) {
-                console.error("Error obteniendo favoritos:", error);
-            }
-        };
-
-        fetchFavorites();
-    }, [authToken, post.id]);
 
     const handleFavoriteClick = async () => {
         if (!authToken) {
@@ -70,12 +57,13 @@ export default function PetCard({ post, className, isPost }: PetCardProps) {
         try {
             if (isFavorite) {
                 await deleteFavorite(post.id, authToken);
+                setFavorites((prev: Favorites[]) => prev.filter((fav) => fav.postId !== post.id));
                 setSuccessMessage("Publicación eliminada de favoritos");
             } else {
                 await addFavorite(post.id, authToken);
+                setFavorites((prev: Favorites[]) => [...prev, { postId: post.id }]);
                 setSuccessMessage("Publicación añadida a favoritos");
             }
-            setIsFavorite(!isFavorite);
         } catch (error) {
             console.error("Error al actualizar favorito", error);
         }
@@ -94,7 +82,7 @@ export default function PetCard({ post, className, isPost }: PetCardProps) {
                 </div>
             )}
             <FavoriteButton variant={isFavorite ? "active" : "desactivated"} // Usa el estado para cambiar el 'variant'
-                onClick={handleFavoriteClick}  className="absolute top-2 right-2 z-10" />
+                onClick={handleFavoriteClick} className="absolute top-2 right-2 z-10" />
             <Link href={isPost ? `/posts/${(post as Post).id}` : `/pets/${(post as Pet).id}`}>
                 <CardImage image={isPost ? (post as Post).urlPhoto : (post as Pet).urlPhoto || ""} />
                 <CardText post={post} />
